+++ cmd2/cmd2.py
@@ -3654,7 +3654,7 @@ class Cmd(cmd.Cmd):
         """
         Disable a command and overwrite its functions
         :param command: the command being disabled
-        :param message_to_print: what to print when this command or its help function is run while disabled
+        :param message_to_print: what to print when this command is run or help is called on it while disabled
         """
         import functools
 
@@ -3683,7 +3683,8 @@ class Cmd(cmd.Cmd):
         """
         Disable an entire category of commands
         :param category: the category to disable
-        :param message_to_print: what to print when anything in this category is run while disabled
+        :param message_to_print: what to print when anything in this category is run or help is called on it
+                                 while disabled
         """
         all_commands = self.get_all_commands()